OK -- maybe this fixes a problem I watched for 3 days.
Don't bg git hooks during clone. This cause(s|d) a race condition in pip
installing from git repos, which tries to rmtree, but the hook was still
running in the background, so pip fails to remove the tags file because
it was getting renamed while the clone was happening.