Attempt to switch back to Telescope.
Tags seems to work now.

The only gap is the <Leader>C mapping, where it seems not
easy to merge two pickers -- specifically a non-recursive
find_files on HOME with a recursive one on .config.

But later if I care enough I'll try to implement a
from-scratch compound picker to do that.