Re-enable tree-sitter highlighting.
Something's changed that was making this line short circuit.

Seems to be two things, one that that variable is true now not 1,
but also that it's nil when this file runs, which probably has to
do with ftplugin loading order changes that I vaguely recall.

No time to look right now, but probably this belongs in after/
now.