Change our diagnostic mapping to be smarter about when to show floats.
It's useless seeing a float that matches the virtual text we already
see.

Obviously we could make this even smarter by checking that virtual text
isn't disabled, or maybe having multilines isn't enough and we need to
check the length and compare to the width of the window, etc., but we
can worry about that when we hit it.