Disable lazy.nvim trying to install from rocks.
I have no idea what changed here, there are commits recently (in the
last day) which in lazy.nvim seem to tweak its luarocks behavior.

It seems like obviously it has begun to have different behavior if it
sees a .rockspec in the plugin's checkout, but I don't know why that's a
good idea or what the expected fix for general users is (for users where
luarocks is not working).

What I do know is that lean.nvim isn't the only plugin suddenly randomly
showing build errors every time neovim opens with this setting on.

nvim-dap-python seems to show the same noise.

So go away for now. We'll see if I have time to investigate further.

Refs: Julian/lean.nvim#432